FIX: Tests are not executed if no testplan supplied
Problem: When executing "dodo <testdirectory>", dod should execute
all the tests but it was not doing that. Instead it didn't execute
even a single test.

Solution: When you supply a testplan, test function names are
extracted and added to a set. This set is declared as a blank set
is even if testplan is not executed testset is a set with length 0.
In Introspector there is a check to see that testplan is None. Here
because our set doesn't go as None ever and so Introspector method
gets a blank set containing no test functions to be executed. As a
solution, set is declared as None and is instantiated only if test
plan is supplied.